# src/database/a7_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'LBAW ERASMUPS Web API'
  description: 'Web Resources Specification (A7) for ERASMUPS'
servers:
  - url: http://lbaw-prod.fe.up.pt
    description: Production server
...
paths:  
  /login:  
      get:
          operationId: R101  
          summary: 'R101: Login Form' 
          description: 'Provide login form. Acess: PUB'  
          tags:  
              - 'M01: Authentication'
          responses:  
              '200':
                  description: 'Ok. Show [UI12](https://git.fe.up.pt/lbaw/lbaw2021/lbaw2164/-/wikis/er#ui12-sign-in)'
      post:  
          operationId: R102  
          summary: 'R102: Login Action'  
          description: 'Processes the login form submission. Acess: PUB'
          tags:  
              - 'M01: Authentication'
          requestBody:
              required: true
              content:
                        application/x-www-form-urlencoded:
                            schema:
                                type: object
                                properties:
                                    username:
                                        type: string
                                    password:
                                        type: string
                                required:
                                   - username
                                   - password
          responses:  
              '302':
                  description: 'Redirect after processing the login credentials.'
                  headers:
                      location:
                          schema:
                              type: string
                          examples:
                              302Success:
                                  description: 'Successful authentication. Redirect to questions.'
                                  value: '/questions'
                              302Error:
                                  description: 'Failed authentication. Redirect to login form.'
                                  value: '/login'
  /login/reset:  
      get:
          operationId: R102  
          summary: 'R102: Request Password Reset Form' 
          description: 'Provide Request Password Reset form. Acess: PUB'  
          tags:  
              - 'M01: Authentication'
          responses:  
              '200':
                  description: 'Ok. Show [UI12](https://git.fe.up.pt/lbaw/lbaw2021/lbaw2164/-/wikis/er#ui12-sign-in)'
      post:  
          operationId: R103  
          summary: 'R103: Request Password Reset Token'  
          description: 'Processes the Request Password Reset and sends email with token. Acess: PUB'
          tags:  
              - 'M01: Authentication'
          requestBody:
              required: true
              content:
                        application/x-www-form-urlencoded:
                            schema:
                                type: object
                                properties:
                                    email:
                                        type: string
                                required:
                                   - email
          responses:  
              '302':
                  description: 'Redirect after processing the request password reset token.'
                  headers:
                      location:
                          schema:
                              type: string
                          examples:
                              302Success:
                                  description: 'Successfully sent email for password recovery. Redirect to Request Password Reset Form.'
                                  value: '/login/reset'
                              302Error:
                                  description: 'Failed to process the password recovery. Redirect to Request Password Reset Form.'
                                  value: '/login/reset'
...
  /questions:  
      get:  
          operationId: R301
          summary: 'R301: Show questions'
          description: 'Show the questions that contain text equals to the one inputed by the user or shows all of them if nothing was inputed. Access: PUB'
          tags:
              - 'M03: Questions, Answers, Comments'
          parameters:
              - in: query
                name: query
                description: String to use for full-text search
                schema:
                    type: string
                required: false
          responses:
              '200':
                  description: 'Showing questions based on user input'
                  content:
                      application/json:
                          schema:
                              type: array
                              questions:
                                  type: object
                                  properties:
                                      title:
                                         type: string
                                      text:
                                         type: string
                                      timestamp:
                                         type: string
                                      score:
                                         type: int
                                      tags:
                                         type: string
                          required: false
                          examples:
                              200Success:
                                  description: 'Successful search.'
                                  title: 'Accomodation Porto'
                                  text: "Hi! I'll be a student at FEUP for the next semester and I was wondering if anyone can give me some information about accomodation? Is there a residence for Erasmus students?"
                                  timestamp: '2021-04-02 14:10:22'
                                  score: 10
                                  tags: 'Porto Accomodation UP Residence'
                              200Empty:
                                  description: 'No valid questions found.'
                                  title: ''
                                  text: ''
                                  timestamp: ''
                                  score: null
                                  tags:
                                  type: ''
...
  /question/{id}:  
      get:  
          operationId: R304
          summary: 'R304: View specific question'
          description: 'Show the content of a specific question(votes, comments and answers). Access: PUB'
          tags:
              - 'M03: Questions, Answers, Comments'
          parameters:
              - in: path
                name: id
                schema:
                    type: integer
                required: true
          responses:
          '200':
              description: 'Ok. Show [UI04](https://git.fe.up.pt/lbaw/lbaw2021/lbaw2164/-/wikis/er#ui04-view-question)'
              
      put:  
          operationId: R305
          summary: 'R305 Close question'
          description: 'Close question so that no more answers can be added. Access: QST'
          tags:
              - 'M03: Questions, Answers, Comments'
          requestBody:
              required: true
              content:
                       application/json:
                           schema:
                               id:
                                  type: int
                               required:
                                   - id
          responses:
              '302':
                  description: 'Redirect after closing question.'
                  headers:
                      Location:
                          schema:
                              type: string
                          examples:
                              302Success:
                                  description: 'Closed sucessfully. Redirect to question page.'
                                  value: '/question/{id}'
                              302Failure:
                                  description: 'Failed to close question. Redirect to question page.'
                                  value: '/question/{id}'

      post:  
          operationId: R306
          summary: 'R306 Disassociate self with question'
          description: 'Disassociate self with question. Access: OWN'
          tags:
              - 'M03: Questions, Answers, Comments'
          requestBody:
              required: true
              content:
                       application/json:
                           schema:
                               id:
                                  type: int
                               required:
                                   - id
          responses:
              '302':
                  description: 'Redirect after Disassocition with question.'
                  headers:
                      Location:
                          schema:
                              type: string
                          examples:
                              302Success:
                                  description: 'Disassocited sucessfully. Redirect to question page.'
                                  value: '/question/{id}'
                              302Failure:
                                  description: 'Failed to disassocite with question, question might be available for deletion. Redirect to question page.'
                                  value: '/question/{id}'

      delete:  
          operationId: R307
          summary: 'R307 Delete question'
          description: 'Delete question. Access: OWN'
          tags:
              - 'M03: Questions, Answers, Comments'
          requestBody:
              required: true
              content:
                       application/json:
                           schema:
                               id:
                                  type: int
                               required:
                                   - id
          responses:
              '302':
                  description: 'Redirect after Deletion of question.'
                  headers:
                      Location:
                          schema:
                              type: string
                          examples:
                              302Success:
                                  description: 'Deleted sucessfully. Redirect to home page.'
                                  value: '/aboutUs'
                              302Failure:
                                  description: 'Failed to Delete question, question is only available for disassocition. Redirect to question page.'
                                  value: '/question/{id}'
...